Index testing
See: index_bar.html and index_wide.html

index_bar.html has a sticky side-bar. The sticky side-bar just has
content from another page in it, as I have yet to give much thought to
what we would want a side-bar to have on the home page. (Keep that in
mind when judging whether or not we want a homepage with a sidebar.

index_wide.html is made up of just the main <div> content on the index
page, which stretches given screen size (to a limit)